utils/Dialogs: TSify
